rubocop

+++ spec/spec_helper.rb
@@ -3,13 +3,15 @@
 require 'pry'
 require 'simplecov'
 
-SimpleCov.start('rails') do
-  profiles.delete(:root_filter)
-  filters.clear
-  add_filter do |src|
-    !(src.filename =~ /^#{SimpleCov.root}/) unless src.filename =~ /\/ez_status\//
+if ENV['COVERAGE']
+  SimpleCov.start('rails') do
+    profiles.delete(:root_filter)
+    filters.clear
+    add_filter do |src|
+      src.filename !~ /^#{SimpleCov.root}/ unless src.filename =~ %r{/ez_status/}
+    end
   end
-end if ENV['COVERAGE']
+end
 
 RSpec.configure do |config|
   config.expect_with :rspec do |expectations|
